Add a second seeder

+++ src/Angor/Shared/Protocol/ProjectInvestmentInfo.cs
@@ -13,7 +13,7 @@ public class ProjectInvestmentInfo
     
     public DateTime StartDate { get; set; }
     public DateTime ExpiryDate { get; set; }
-    public double TargetAmount { get; set; }
+    public decimal TargetAmount { get; set; }
     public List<Seeders> Seeders { get; set; } = new();
     public List<Stage> Stages { get; set; } = new();
 }
